requirements.txt
Django==6.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
reportlab>=4.0.7
openpyxl>=3.1.2
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
Pillow>=10.4.0
qrcode[pil]>=7.4.2
python-barcode>=0.15.1
django-redis>=5.4.0
django-ratelimit>=4.1.0
django-extensions>=3.2.3
celery>=5.3.4
redis>=5.0.1

